re-add bundle
