+++ .gitignore
@@ -3,3 +3,4 @@ dist/
 node_modules/
 NetScriptDefinitions.d.ts
 NetscriptDefinitions.d.ts
+bundle/
