+++ package.json
@@ -17,15 +17,18 @@
   "author": "jamesptanner",
   "license": "ISC",
   "devDependencies": {
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-typescript": "^11.1.6",
     "@types/react": "^18.2.70",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^5.35.1",
     "@typescript-eslint/parser": "^5.35.1",
     "bitburner-filesync": "^1.1.5",
+    "concurrently": "^8.2.2",
     "eslint": "^8.22.0",
     "eslint-watch": "^8.0.0",
-    "concurrently": "^8.2.2",
-    "typescript": "^4.7.4",
-    "sync-directory": "^5.1.7"
+    "rollup": "^4.13.2",
+    "sync-directory": "^5.1.7",
+    "typescript": "^4.7.4"
   }
 }
